backend/requirements.txt
fastapi[all]>=0.111.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
python-gitlab>=4.4.0
toml>=0.10.2
uvicorn>=0.29.0
